# docs/openapi.yml
openapi: 3.0.1
info:
  title: Opdex Platform API
  version: 1.0.0
  contact:
    name: Opdex Contributors
    url: https://github.com/Opdex/opdex-v1-api
  description: The Opdex Platform Web API is an interface for querying information from the Opdex contract indexer and quoting Opdex smart contract transactions. The Opdex contract indexer is a job that continually runs and stores relevant broadcast transaction data, in a way that represents the current known state of the Opdex protocol at the present time.
  license:
    name: MIT
    url: https://github.com/Opdex/opdex-v1-api/blob/main/LICENSE
servers:
  - url: https://test-api.opdex.com
...
components:
  securitySchemes:
    opdexAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authorize with a bearer token obtained from SignalR
  responses:
...
    TooManyRequests:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 30
          description: Indicates how many seconds to wait before making a follow-up request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/problemDetails"
          example:
            type: https://httpstatuses.com/429
            title: Too Many Requests
            detail: "Quota exceeded. Maximum allowed: 500 per 60s. Please try again in 30 second(s)."
            status: 429
            extensions:
              traceId: 00-00000000000000000000000000000000-0000000000000000-00
    InternalServerError:
      description: Unexpected error occurred
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/problemDetails"
          example:
            type: https://httpstatuses.com/500
            title: Internal Server Error
            status: 500
            extensions:
              traceId: 00-00000000000000000000000000000000-0000000000000000-00
  schemas:
    problemDetails:
      type: object
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem type
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server for this occurrence of the problem
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem
        extensions:
          type: object
          additionalProperties:
            nullable: true
          description: Non-standard problem details extensions that may be ignored
      description: Carries machine-readable details of errors in a HTTP response
...
    sha256:
      type: string
      description: SHA256 hash value
      pattern: ^[A-Fa-f0-9]{64}$
      minLength: 64
      maxLength: 64
...
    blockResponse:
      type: object
      description: Details about a block
      properties:
        hash:
          $ref: "#/components/schemas/sha256"
          description: SHA-256 hash of the block
        height:
          type: integer
          format: uint64
          minimum: 1
          description: Block number
        time:
          type: string
          format: date-time
          description: Block timestamp as determined by the block producer
        medianTime:
          type: string
          format: date-time
          description: Network-adjusted timestamp
...
    indexerStatusResponse:
      type: object
      description: Status of the Opdex transaction indexer
      properties:
        latestBlock:
          $ref: "#/components/schemas/blockResponse"
          description: Details of the latest block indexed
        available:
          type: boolean
          description: True if the indexer process is active, otherwise false
        locked:
          type: boolean
          description: True if the indexer process is currently running, otherwise false
        instanceId:
          type: string
          description: Unique identifier for the API instance that locked the indexer
        reason:
          type: string
          enum:
            - Deploying
            - Indexing
            - Rewinding
          description: Reason that the indexer process is currently running
        modifiedDate:
          type: string
          format: date-time
          description: When the indexer status last changed
...
paths:
...
  /indexer:
    get:
      tags:
        - Indexer
      summary: Get Indexer Status
      description: Retrieves status details for the Opdex transaction indexer.
      operationId: getIndexerStatus
      responses:
        200:
          description: Indexer status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/indexerStatusResponse"
              example:
                latestBlock:
                  hash: 05179f49edccc26387a1d9200e2525ca64067b294d31966450e111d983ddefc1
                  height: 2987787
                  time: 2022-01-13T13:09:04Z
                  medianTime: 2022-01-13T13:06:56Z
                available: true
                locked: true
                instanceId: bc399269-74cb-4147-a7c3-f2d296fa4a9b
                reason: Indexing
                modifiedDate: 2022-01-13T13:09:56Z
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
...